+++ octopy/rest/rest_delete.py
